ANZARD-61: downcast cross question validation set and constant params
Downcasts the Cross Question Validation constant, set, conditional constant and conditional set params on object save. This facilitates a case insensitive comparison of choice question answers throughout the application, in this case when performing some kind of cross question validation of the choice answer against a pre-defined set or constant.